Split region_job.py into materialized and virtual modules (PR 3c)
Move-only refactor: pull MaterializedRegionJob into
common/materialized_region_job.py and VirtualRef/VirtualRegionJob (plus
the _exists_many helper) into common/virtual_region_job.py, out of the
shared common/region_job.py. The base RegionJob, SourceFileCoord,
SourceFileResult, and SourceFileStatus stay in region_job.py.

The moved class bodies are byte-identical to the originals; only imports
and importer references change. The download-error logging tests follow
the logger name to its new module.

RegionJob.process stays a base-class stub for now; revisiting whether it
can live on MaterializedRegionJob is deliberately a separate change.
