+++ src/reformatters/common/region_job.py
@@ -1,36 +1,21 @@
-import asyncio
-import concurrent.futures
-import os
-from collections.abc import Callable, Iterator, Mapping, Sequence
-from concurrent.futures import ProcessPoolExecutor, ThreadPoolExecutor
-from contextlib import suppress
-from copy import deepcopy
+from collections.abc import Callable, Mapping, Sequence
 from enum import Enum, auto
-from http import HTTPStatus
 from itertools import chain, pairwise
-from multiprocessing.shared_memory import SharedMemory
 from pathlib import Path
 from typing import (
     Annotated,
     Any,
     ClassVar,
-    Final,
     Generic,
-    NamedTuple,
     Self,
     TypeVar,
     get_args,
 )
 
-import httpx
-import icechunk
 import numpy as np
 import pandas as pd
 import pydantic
 import xarray as xr
-import zarr
-from icechunk import VirtualChunkSpec
-from icechunk.store import IcechunkStore
 from pydantic import (
     AfterValidator,
     Field,
@@ -39,27 +24,18 @@ from pydantic import (
     field_validator,
 )
 from zarr.abc.store import Store
-from zarr.core.metadata import ArrayV3Metadata
 
-from reformatters.common import storage, template_utils
-from reformatters.common.binary_rounding import round_float32_inplace
+from reformatters.common import storage
 from reformatters.common.config_models import DataVar
 from reformatters.common.iterating import dimension_slices, split_groups
 from reformatters.common.logging import get_logger
-from reformatters.common.pydantic import FrozenBaseModel, replace
-from reformatters.common.shared_memory_utils import (
-    create_data_array_and_template,
-    make_shared_buffer,
-    write_shards,
-)
+from reformatters.common.pydantic import FrozenBaseModel
 from reformatters.common.types import (
     AppendDim,
-    ArrayND,
     DatetimeLike,
     Dim,
     Timestamp,
 )
-from reformatters.common.zarr import copy_data_var
 
 log = get_logger(__name__)
 
@@ -473,653 +449,6 @@ class RegionJob(pydantic.BaseModel, Generic[DATA_VAR, SOURCE_FILE_COORD]):
 
     def __repr__(self) -> str:
         return f"{type(self).__name__}(region=({self.region.start}, {self.region.stop}), data_vars={[v.name for v in self.data_vars]})"
-
-
-class MaterializedRegionJob(
-    RegionJob[DATA_VAR, SOURCE_FILE_COORD], Generic[DATA_VAR, SOURCE_FILE_COORD]
-):
-    # Reformats source data into rechunked (materialized) chunk data: download
-    # source files, read them into shared-memory arrays, write output shards.
-    # Base class for all existing rechunked datasets.
-
-    # Limit the number of variables processed in each download group if set.
-    # If value is less than len(data_vars), downloading, reading/recompressing, and writing steps
-    # will be pipelined within a region job.
-    max_vars_per_download_group: ClassVar[int | None] = None
-
-    # Subclasses can override this to control download parallelism
-    # This particularly useful of the data source cannot handle a large number of concurrent requests
-    download_parallelism: int = (os.cpu_count() or 1) * 2
-
-    # Subclasses can override this to control read parallelism.
-    # This is useful in cases where many threads reading data into shared memory
-    # causes deadlocks or resource contention.
-    read_parallelism: int = os.cpu_count() or 1
-
-    def download_file(self, coord: SOURCE_FILE_COORD) -> Path:
-        """Download the file for the given coordinate and return the local path."""
-        raise NotImplementedError(
-            "Download the file for the given coordinate and return the local path."
-        )
-
-    def read_data(
-        self,
-        coord: SOURCE_FILE_COORD,
-        data_var: DATA_VAR,
-    ) -> ArrayND[np.generic]:
-        """
-        Read and return the data chunk for the given variable and source file coordinate.
-
-        Subclasses must implement this to load the data (e.g., from a file or remote source)
-        for the specified coord and data_var. The returned array will be written into the shared
-        output array by the base class.
-
-        Parameters
-        ----------
-        coord : SOURCE_FILE_COORD
-            The coordinate specifying which file and region to read.
-        data_var : DATA_VAR
-            The data variable metadata.
-
-        Returns
-        -------
-        ArrayND[np.generic]
-            The loaded data.
-        """
-        raise NotImplementedError(
-            "Read and return data for the given variable and source file coordinate."
-        )
-
-    def apply_data_transformations(
-        self, data_array: xr.DataArray, data_var: DATA_VAR
-    ) -> None:
-        """
-        Apply in-place data transformations to the output data array for a given data variable.
-
-        This method is called after reading all data for a variable into the shared-memory array,
-        and before writing shards to the output store. The default implementation applies binary
-        rounding to float32 arrays if `data_var.internal_attrs.keep_mantissa_bits` is set.
-
-        Subclasses may override this method to implement additional transformations such as
-        deaccumulation, interpolation or other custom logic. All transformations should be
-        performed in-place (don't copy `data_array`, it's large).
-
-        Parameters
-        ----------
-        data_array : xr.DataArray
-            The output data array to be transformed in-place.
-        data_var : DATA_VAR
-            The data variable metadata object, which may contain transformation parameters.
-        """
-        keep_mantissa_bits = data_var.internal_attrs.keep_mantissa_bits
-        if isinstance(keep_mantissa_bits, int):
-            round_float32_inplace(
-                data_array.values, keep_mantissa_bits=keep_mantissa_bits
-            )
-
-    @classmethod
-    def process_worker_jobs(
-        cls,
-        worker_jobs: Sequence[RegionJob[DATA_VAR, SOURCE_FILE_COORD]],
-        store_factory: storage.StoreFactory,
-        branch_name: str,
-        worker_index: int,
-    ) -> dict[str, list[SourceFileResult]]:
-        """Write all of this worker's jobs to ``branch_name`` in a single commit."""
-        # One commit per worker; an empty job set would make an empty icechunk
-        # commit (which raises), so the caller must filter empty workers out.
-        assert worker_jobs, "process_worker_jobs requires at least one job"
-        primary_store = store_factory.primary_store(writable=True, branch=branch_name)
-        replica_stores = store_factory.replica_stores(writable=True, branch=branch_name)
-
-        worker_results: dict[str, list[SourceFileResult]] = {}
-        for job in worker_jobs:
-            template_utils.write_metadata(job.template_ds, job.tmp_store)
-            results = job.process(
-                primary_store=primary_store, replica_stores=replica_stores
-            )
-            for var_name, coords in results.items():
-                worker_results.setdefault(var_name, []).extend(
-                    SourceFileResult(
-                        status=c.status,
-                        out_loc={**c.out_loc()},
-                        url=c.get_url(),
-                    )
-                    for c in coords
-                )
-
-        now = pd.Timestamp.now(tz="UTC")
-        storage.commit_if_icechunk(
-            f"Update worker {worker_index} at {now.strftime('%Y-%m-%dT%H:%M:%SZ')}",
-            primary_store,
-            replica_stores,
-        )
-        return worker_results
-
-    def process(
-        self,
-        primary_store: Store,
-        replica_stores: list[Store],
-    ) -> Mapping[str, Sequence[SOURCE_FILE_COORD]]:
-        """
-        Orchestrate the full region job processing pipeline.
-
-        1. Group data variables for efficient processing (e.g., by file type or batch size)
-        2. Write zarr metadata to tmp store for region="auto" support
-        3. For each group of data variables:
-            a. Download all required source files
-            b. For each variable in the group:
-                i.   Read data from source files into the shared array
-                ii.  Apply any required data transformations (e.g., rounding, deaccumulation)
-                iii. Write output shards to the tmp_store
-                iv.  Upload chunk data from tmp_store to the primary store
-
-        Returns
-        -------
-        Mapping[str, Sequence[SOURCE_FILE_COORD]]
-            Mapping from variable names to their source file coordinates with final processing status.
-        """
-        processing_region_ds, output_region_ds = self._get_region_datasets()
-
-        data_var_groups = self.source_groups(self.data_vars)
-        if self.max_vars_per_download_group is not None:
-            data_var_groups = split_groups(
-                data_var_groups, self.max_vars_per_download_group
-            )
-
-        results: dict[str, Sequence[SOURCE_FILE_COORD]] = {}
-        upload_futures: list[Any] = []
-
-        with (
-            make_shared_buffer(processing_region_ds) as shared_buffer,
-            ThreadPoolExecutor(max_workers=1) as download_executor,
-            ThreadPoolExecutor(max_workers=1) as upload_executor,
-            ProcessPoolExecutor(max_workers=os.cpu_count() or 1) as write_executor,
-        ):
-            log.info(f"Starting {self!r}")
-
-            # Submit all download tasks to the executor
-            download_futures = {}
-            for data_var_group in data_var_groups:
-                data_var_names = [v.name for v in data_var_group]
-                source_file_coords = self.generate_source_file_coords(
-                    processing_region_ds, data_var_group
-                )
-                download_future = download_executor.submit(
-                    self._download_processing_group, source_file_coords, data_var_names
-                )
-                download_futures[download_future] = data_var_group
-
-            # Process downloaded data var groups as they complete
-            for download_future in concurrent.futures.as_completed(download_futures):
-                data_var_group = download_futures[download_future]
-                source_file_coords = download_future.result()
-                log.info(f"Downloaded: {[v.name for v in data_var_group]}")
-
-                # Process one data variable at a time to ensure a single user of
-                # the shared buffer at a time and to reduce peak memory usage
-                for data_var in data_var_group:
-                    # Copy so we have a unique status per variable, not per variable group
-                    data_var_source_file_coords = deepcopy(source_file_coords)
-                    data_array, data_array_template = create_data_array_and_template(
-                        processing_region_ds,
-                        data_var.name,
-                        shared_buffer,
-                        fill_value=data_var.encoding.fill_value,
-                    )
-                    data_var_source_file_coords = self._read_into_data_array(
-                        data_array,
-                        data_var,
-                        data_var_source_file_coords,
-                    )
-                    self.apply_data_transformations(
-                        data_array,
-                        data_var,
-                    )
-                    self._write_shards(
-                        data_array_template,
-                        shared_buffer,
-                        output_region_ds,
-                        self.tmp_store,
-                        write_executor,
-                    )
-
-                    upload_futures.append(
-                        upload_executor.submit(
-                            copy_data_var,
-                            data_var.name,
-                            self.region,
-                            self.template_ds,
-                            self.append_dim,
-                            self.tmp_store,
-                            primary_store,
-                            replica_stores=replica_stores,
-                        )
-                    )
-
-                    results[data_var.name] = data_var_source_file_coords
-
-                self._cleanup_local_files(source_file_coords)  # after _group_ is done
-
-        for future in concurrent.futures.as_completed(upload_futures):
-            if (e := future.exception()) is not None:
-                raise e
-
-        return results
-
-    def _get_region_datasets(self) -> tuple[xr.Dataset, xr.Dataset]:
-        ds: xr.Dataset = self.template_ds[[v.name for v in self.data_vars]]  # ty: ignore[invalid-assignment]
-        processing_region = self.get_processing_region()
-        processing_region_ds = ds.isel({self.append_dim: processing_region})
-        output_region_ds = ds.isel({self.append_dim: self.region})
-        return processing_region_ds, output_region_ds
-
-    def _download_processing_group(
-        self,
-        source_file_coords: Sequence[SOURCE_FILE_COORD],
-        data_var_names: Sequence[str],
-    ) -> list[SOURCE_FILE_COORD]:
-        """
-        Download specified source files in parallel.
-
-        Returns
-        -------
-        list[SOURCE_FILE_COORD]
-            List of SourceFileCoord objects with updated download status and path.
-        """
-
-        def _call_download_file(coord: SOURCE_FILE_COORD) -> SOURCE_FILE_COORD:
-            try:
-                path = self.download_file(coord)
-                return replace(coord, downloaded_path=path)
-            except Exception as e:
-                updated_coord = replace(coord, status=SourceFileStatus.DownloadFailed)
-
-                # For recent files, we expect some files to not exist yet, just log the path
-                # else, log exception so it is caught by error reporting but doesn't stop processing
-                append_dim_coord = coord.append_dim_coord
-                two_days_ago = pd.Timestamp.now() - pd.Timedelta(hours=48)
-                is_not_found = isinstance(e, FileNotFoundError) or (
-                    isinstance(e, httpx.HTTPStatusError)
-                    and e.response.status_code
-                    in (HTTPStatus.FORBIDDEN, HTTPStatus.NOT_FOUND)
-                )
-                if (
-                    is_not_found
-                    and isinstance(append_dim_coord, np.datetime64 | pd.Timestamp)
-                    and append_dim_coord > two_days_ago
-                ):
-                    log.info(" ".join(str(e).split("\n")[:2]))
-                else:
-                    log.exception(f"Download failed {coord.get_url()}")
-
-                return updated_coord
-
-        log.info(f"Downloading {data_var_names} in {len(source_file_coords)} files...")
-        with ThreadPoolExecutor(
-            max_workers=self.download_parallelism
-        ) as download_executor:
-            return list(download_executor.map(_call_download_file, source_file_coords))
-
-    def _read_into_data_array(
-        self,
-        out: xr.DataArray,
-        data_var: DATA_VAR,
-        source_file_coords: Sequence[SOURCE_FILE_COORD],
-    ) -> list[SOURCE_FILE_COORD]:
-        """
-        Reads data from source files into `out`.
-        Returns a list of coords with the final status.
-        """
-        # Skip coords where the download failed
-        read_coords = (
-            c for c in source_file_coords if c.status == SourceFileStatus.Processing
-        )
-
-        def _read_and_write_one(coord: SOURCE_FILE_COORD) -> ArrayND[np.generic]:
-            return self.read_data(coord, data_var)
-
-        # Index is used to maintain order of coords
-        updated_coords: dict[int, SOURCE_FILE_COORD] = {}
-
-        log.info(f"Reading {data_var.name}...")
-        with ThreadPoolExecutor(max_workers=self.read_parallelism) as executor:
-            futures = {}
-            for i, coord in enumerate(read_coords):
-                future = executor.submit(_read_and_write_one, coord)
-                futures[future] = (i, coord)
-
-            for future in concurrent.futures.as_completed(futures):
-                index, coord = futures[future]
-                try:
-                    out.loc[coord.out_loc()] = future.result()
-                    updated_coords[index] = replace(
-                        coord, status=SourceFileStatus.Succeeded
-                    )
-                except Exception:
-                    log.exception(f"Read failed {coord.downloaded_path}")
-                    updated_coords[index] = replace(
-                        coord, status=SourceFileStatus.ReadFailed
-                    )
-                finally:
-                    # as_completed retains futures and results; clear to avoid ~2x peak memory
-                    future._result = None  # noqa: SLF001
-
-        sorted_updated_coords = [updated_coords[i] for i in sorted(updated_coords)]
-        return sorted_updated_coords
-
-    def _write_shards(
-        self,
-        processing_region_da_template: xr.DataArray,
-        shared_buffer: SharedMemory,
-        output_region_ds: xr.Dataset,
-        tmp_store: Path,
-        write_executor: ProcessPoolExecutor,
-    ) -> None:
-        write_shards(
-            processing_region_da_template,
-            shared_buffer,
-            self.append_dim,
-            output_region_ds,
-            tmp_store,
-            write_executor,
-        )
-
-    def _cleanup_local_files(
-        self, source_file_coords: Sequence[SOURCE_FILE_COORD]
-    ) -> None:
-        for coord in source_file_coords:
-            if coord.downloaded_path:
-                with suppress(FileNotFoundError):
-                    coord.downloaded_path.unlink()
-
-
-class VirtualRef(NamedTuple):
-    """A single virtual chunk reference: one GRIB message -> one zarr chunk.
-    `out_loc` is the output cell the message fills (e.g. {init_time, lead_time}).
-    """
-
-    data_var: DataVar[Any]
-    out_loc: Mapping[Dim, CoordinateValue]
-    location: str
-    offset: int
-    length: int
-
-
-class VirtualRegionJob(
-    RegionJob[DATA_VAR, SOURCE_FILE_COORD], Generic[DATA_VAR, SOURCE_FILE_COORD]
-):
-    """Base class for processing a region of virtual Icechunk datasets that point to external files."""
-
-    # Locked to None: an integer max_vars_per_job would split one source file's
-    # variables across separate jobs that commit independently, breaking the
-    # per-file commit atomicity virtual readers rely on.
-    max_vars_per_job: ClassVar[Final[int | None]] = None
-
-    def process_virtual_refs(
-        self,
-        remaining: Sequence[SOURCE_FILE_COORD],
-    ) -> Iterator[Sequence[VirtualRef]]:
-        """Discover available source files among `remaining` and yield their virtual references.
-
-        Each yield is one commit's worth of refs. The contract is that a yield
-        holds *whole* source files — every ref of a file is emitted in the same
-        batch, never split across yields — so a reader on `main` sees all of a
-        file's data or none of it. This is a discipline the generator must keep
-        (and tests check). The generator owns the batching policy: backfill yields
-        batches of ~N whole files to amortize commit overhead; operational yields
-        ~1 file at a time for per-file visibility.
-        """
-        raise NotImplementedError(
-            "Yield batches of VirtualRefs, each one commit's worth of whole source files."
-        )
-
-    def representative_var(self, coord: SOURCE_FILE_COORD) -> DATA_VAR:  # noqa: ARG002
-        """The variable whose chunk presence means `coord`'s file is fully ingested.
-
-        Used by filter_already_present to probe a single representative cell per
-        file. Default: the first instant var (most likely to have data at every
-        step), else the first data var; valid when every file contains every
-        variable. Override for one-variable-per-file packings (e.g. GEFS v12
-        reforecast) to return the candidate's own variable.
-        """
-        return next(
-            (var for var in self.data_vars if var.attrs.step_type == "instant"),
-            self.data_vars[0],
-        )
-
-    def filter_already_present(
-        self,
-        candidates: Sequence[SOURCE_FILE_COORD],
-        store: IcechunkStore,
-    ) -> list[SOURCE_FILE_COORD]:
-        """Drop candidates whose representative chunk is already in the manifest.
-
-        Probes ref existence (store.exists) - never reads or decodes a chunk, which
-        would trigger a download + decode.
-        """
-        group = zarr.open_group(store, mode="r")
-        keyed: list[tuple[SOURCE_FILE_COORD, str | None]] = []
-        for coord in candidates:
-            var = self.representative_var(coord)
-            index = self.chunk_key(coord.out_loc(), var)
-            if index is None:
-                keyed.append((coord, None))
-                continue
-            array = group[var.name]
-            assert isinstance(array, zarr.Array)
-            metadata = array.metadata
-            assert isinstance(metadata, ArrayV3Metadata)
-            key = f"{array.path}/{metadata.chunk_key_encoding.encode_chunk_key(index)}"
-            keyed.append((coord, key))
-
-        present = _exists_many(store, [key for _, key in keyed if key is not None])
-        return [coord for coord, key in keyed if key is None or not present[key]]
-
-    # ----- Most subclasses will not need to override the attributes and methods below -----
-
-    @classmethod
-    def process_worker_jobs(
-        cls,
-        worker_jobs: Sequence[RegionJob[DATA_VAR, SOURCE_FILE_COORD]],
-        store_factory: storage.StoreFactory,
-        branch_name: str,
-        worker_index: int,  # noqa: ARG003 - virtual jobs commit per batch with their own messages
-    ) -> dict[str, list[SourceFileResult]]:
-        """Drive each job's per-batch virtual write loop on ``branch_name``.
-
-        Always returns an empty dict of results: virtual refs live in the icechunk
-        manifest, so there is nothing to thread back to finalize.
-        """
-        assert worker_jobs, "process_worker_jobs requires at least one job"
-        primary_repo, replica_repos = store_factory.icechunk_primary_and_replica_repos()
-        for job in worker_jobs:
-            assert isinstance(job, VirtualRegionJob)
-            job.process_virtual(primary_repo, replica_repos, branch_name)
-        return {}
-
-    def process_virtual(
-        self,
-        primary_repo: icechunk.Repository,
-        replica_repos: Sequence[icechunk.Repository],
-        branch: str,
-    ) -> None:
-        """Run the virtual write loop, committing each yielded batch atomically.
-
-        Same loop for backfill and operational; the only difference is the
-        `branch` the driver opened (a temp branch for backfill, "main" for the
-        single-writer operational update). sync_dims_to grows the store lazily (a
-        no-op on the pre-sized backfill branch). A fresh writable session is opened
-        per batch because an icechunk session becomes read-only after commit.
-        """
-        readonly_store = primary_repo.readonly_session(branch).store
-        candidates = self.generate_source_file_coords(
-            self._processing_region_ds(), self.data_vars
-        )
-        remaining = self.filter_already_present(candidates, readonly_store)
-        current_size = self._append_dim_size(readonly_store)
-
-        for refs in self.process_virtual_refs(remaining):
-            # The generator yields whole files, so a batch always has refs; emitting
-            # them always dirties the session, so the commit below is never empty
-            # (an empty icechunk commit raises rather than no-ops).
-            assert refs, "process_virtual_refs yielded an empty batch"
-            primary_session = primary_repo.writable_session(branch)
-            replica_sessions = [repo.writable_session(branch) for repo in replica_repos]
-            stores = [primary_session.store, *(s.store for s in replica_sessions)]
-
-            needed_size = self._needed_append_dim_size(refs)
-            if needed_size > current_size:
-                self.sync_dims_to(stores, needed_size)
-                current_size = needed_size
-
-            self._emit_refs(stores, refs)
-
-            now = pd.Timestamp.now(tz="UTC")
-            storage.commit_if_icechunk(
-                f"Update at {now.strftime('%Y-%m-%dT%H:%M:%SZ')}",
-                primary_session.store,
-                [s.store for s in replica_sessions],
-            )
-
-    def _emit_refs(
-        self, stores: Sequence[IcechunkStore], refs: Sequence[VirtualRef]
-    ) -> None:
-        specs_by_var: dict[str, list[VirtualChunkSpec]] = {}
-        for ref in refs:
-            index = self.chunk_key(ref.out_loc, ref.data_var)
-            assert index is not None, (
-                f"ref {ref.data_var.name} {dict(ref.out_loc)} did not resolve to a "
-                "chunk index after expansion"
-            )
-            specs_by_var.setdefault(ref.data_var.name, []).append(
-                VirtualChunkSpec(
-                    index=list(index),
-                    location=ref.location,
-                    offset=ref.offset,
-                    length=ref.length,
-                )
-            )
-        for store in stores:
-            for var_name, specs in specs_by_var.items():
-                failed = store.set_virtual_refs(
-                    var_name, specs, validate_containers=True
-                )
-                assert failed is None, (
-                    f"{len(failed)} virtual ref(s) for {var_name} did not match a "
-                    f"registered container, e.g. {failed[:3]}"
-                )
-
-    def chunk_key(
-        self, out_loc: Mapping[Dim, CoordinateValue], var: DataVar[Any]
-    ) -> tuple[int, ...] | None:
-        """Map a source message's coordinate labels to its zarr chunk index.
-
-        Returns None if a label is not in the template's coords, which the filter
-        treats as "remaining". Asserts each label lands on an exact chunk boundary -
-        a virtual chunk is exactly one GRIB message, so any non-aligned position is
-        a bug.
-        """
-        # Geometry (dim order, chunk sizes, coord positions) is read from the
-        # checked-in template (the authoritative metadata, git-reviewable), not the
-        # in-code DataVar.encoding which could drift from it. So the index matches
-        # what readers resolve and does not depend on how far the store has been
-        # expanded. Shared by the filter and the emitter so they cannot disagree.
-        template_var = self.template_ds[var.name]
-        dims = tuple(str(d) for d in template_var.dims)
-        chunks = template_var.encoding["chunks"]
-        assert len(chunks) == len(dims)
-        labels = {str(dim): label for dim, label in out_loc.items()}
-
-        index: list[int] = []
-        for dim, chunk_size in zip(dims, chunks, strict=True):
-            if dim in labels:
-                position = int(
-                    self.template_ds.get_index(dim).get_indexer(
-                        pd.Index([labels[dim]])
-                    )[0]
-                )
-                if position < 0:
-                    return None  # label not in coords -> not yet a position in this dataset
-            else:
-                # A dim absent from out_loc (e.g. lat/lon) must be a single chunk;
-                # otherwise every ref would collapse to chunk 0 along it (later refs
-                # overwriting earlier, the rest left as permanent fill-value holes).
-                # A multi-chunk dim (e.g. ensemble_member packed one-per-file) must
-                # appear in out_loc so it resolves to the right chunk.
-                assert chunk_size >= template_var.sizes[dim], (
-                    f"dim {dim} is absent from out_loc but spans multiple chunks "
-                    f"(size {template_var.sizes[dim]}, chunk {chunk_size}); out_loc "
-                    "must locate every multi-chunk dim."
-                )
-                position = 0
-            chunk_index, remainder = divmod(position, chunk_size)
-            assert remainder == 0, (
-                f"{dim}={labels.get(dim)} maps to position {position}, which is not "
-                f"on a chunk boundary (chunk size {chunk_size}); a virtual chunk must "
-                "be exactly one GRIB message."
-            )
-            index.append(chunk_index)
-        return tuple(index)
-
-    def sync_dims_to(
-        self, stores: Sequence[IcechunkStore], needed_append_dim_size: int
-    ) -> None:
-        """Grow the append dim (and its dependent coords) to `needed_append_dim_size`."""
-        # Writes only the new positions by appending the corresponding slice of the
-        # already-derived template; existing positions are untouched. Data vars stay
-        # dask-lazy under compute=False, so the decode-only serializer is never
-        # invoked. A no-op for any store already covering the size (so backfill's
-        # pre-sized branch never resizes, and parallel workers never conflict). Each
-        # store's missing slice is computed from its own committed size: replicas
-        # commit before the primary, so a partial commit can leave a replica ahead,
-        # and appending the primary's slice to it would duplicate positions.
-        for store in stores:
-            current_size = self._append_dim_size(store)
-            if needed_append_dim_size <= current_size:
-                continue
-            slice_ds = self.template_ds.isel(
-                {self.append_dim: slice(current_size, needed_append_dim_size)}
-            )
-            slice_ds.to_zarr(
-                store, append_dim=self.append_dim, compute=False, consolidated=False
-            )
-
-    def _needed_append_dim_size(self, refs: Sequence[VirtualRef]) -> int:
-        positions = self.template_ds.get_index(self.append_dim).get_indexer(
-            pd.Index([ref.out_loc[self.append_dim] for ref in refs])
-        )
-        assert (positions >= 0).all(), (
-            "a ref's append-dim label is not present in the template"
-        )
-        return int(positions.max()) + 1
-
-    def _append_dim_size(self, store: IcechunkStore) -> int:
-        array = zarr.open_group(store, mode="r")[self.append_dim]
-        assert isinstance(array, zarr.Array)
-        return int(array.shape[0])
-
-    def _processing_region_ds(self) -> xr.Dataset:
-        processing_region = self.get_processing_region()
-        # Virtual refs point at raw source bytes, so no surrounding buffer is ever
-        # needed; a buffered region would make adjacent backfill workers generate
-        # overlapping candidates and emit refs into each other's regions.
-        assert processing_region == self.region, (
-            "VirtualRegionJob.get_processing_region must equal self.region"
-        )
-        ds: xr.Dataset = self.template_ds[[v.name for v in self.data_vars]]  # ty: ignore[invalid-assignment]
-        return ds.isel({self.append_dim: processing_region})
-
-
-def _exists_many(store: IcechunkStore, keys: Sequence[str]) -> dict[str, bool]:
-    """Probe many chunk keys concurrently (store.exists is async)."""
-    if not keys:
-        return {}
-
-    async def _check() -> list[bool]:
-        return list(await asyncio.gather(*(store.exists(key) for key in keys)))
-
-    return dict(zip(keys, asyncio.run(_check()), strict=True))
 
 
 class SourceFileResult(FrozenBaseModel):
