+++ src/reformatters/contrib/noaa/ndvi_cdr/analysis/region_job.py
@@ -20,9 +20,11 @@ from reformatters.common.download import (
 )
 from reformatters.common.iterating import item
 from reformatters.common.logging import get_logger
+from reformatters.common.materialized_region_job import (
+    MaterializedRegionJob,
+)
 from reformatters.common.region_job import (
     CoordinateValue,
-    MaterializedRegionJob,
     RegionJob,
     SourceFileCoord,
 )
