+++ src/reformatters/ecmwf/ifs_ens/forecast_15_day_0_25_degree/region_job.py
@@ -14,7 +14,8 @@ from reformatters.common.deaccumulation import deaccumulate_to_rates_inplace
 from reformatters.common.download import http_download_to_disk
 from reformatters.common.iterating import digest, item
 from reformatters.common.logging import get_logger
-from reformatters.common.region_job import MaterializedRegionJob, RegionJob
+from reformatters.common.materialized_region_job import MaterializedRegionJob
+from reformatters.common.region_job import RegionJob
 from reformatters.common.types import (
     AppendDim,
     ArrayFloat32,
