+++ src/reformatters/noaa/gefs/forecast_35_day/region_job.py
@@ -11,7 +11,8 @@ from reformatters.common.binary_rounding import round_float32_inplace
 from reformatters.common.deaccumulation import deaccumulate_to_rates_inplace
 from reformatters.common.iterating import item
 from reformatters.common.logging import get_logger
-from reformatters.common.region_job import MaterializedRegionJob, RegionJob
+from reformatters.common.materialized_region_job import MaterializedRegionJob
+from reformatters.common.region_job import RegionJob
 from reformatters.common.types import AppendDim, ArrayND, DatetimeLike
 from reformatters.noaa.gefs.gefs_config_models import (
     GEFSDataVar,
