+++ src/reformatters/noaa/hrrr/region_job.py
@@ -14,9 +14,11 @@ from reformatters.common.deaccumulation import deaccumulate_to_rates_inplace
 from reformatters.common.download import http_download_to_disk, httpx_download_to_disk
 from reformatters.common.iterating import digest, group_by, item
 from reformatters.common.logging import get_logger
+from reformatters.common.materialized_region_job import (
+    MaterializedRegionJob,
+)
 from reformatters.common.region_job import (
     CoordinateValue,
-    MaterializedRegionJob,
     RegionJob,
     SourceFileCoord,
 )
