+++ tests/common/test_parallel_writes.py
@@ -33,8 +33,10 @@ from reformatters.common.config_models import (
 )
 from reformatters.common.dynamical_dataset import DynamicalDataset
 from reformatters.common.kubernetes import CronJob, ReformatCronJob, ValidationCronJob
-from reformatters.common.region_job import (
+from reformatters.common.materialized_region_job import (
     MaterializedRegionJob,
+)
+from reformatters.common.region_job import (
     SourceFileCoord,
 )
 from reformatters.common.storage import DatasetFormat, StorageConfig
